# Desktop/lumen/lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Learning Management System API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /courses/{courseId}/lessons:
    get:
      operationId: listLessonsForCourse
      tags: [lessons]
      parameters:
        - name: courseId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Lesson"
    post:
      operationId: createLesson
      tags: [lessons]
      parameters:
        - name: courseId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateLessonInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Lesson"
  /lessons/{lessonId}:
    get:
      operationId: getLesson
      tags: [lessons]
      parameters:
        - name: lessonId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Lesson"
    patch:
      operationId: updateLesson
      tags: [lessons]
      parameters:
        - name: lessonId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateLessonInput"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Lesson"
    delete:
      operationId: deleteLesson
      tags: [lessons]
      parameters:
        - name: lessonId
          in: path
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Deleted
  /lessons/{lessonId}/complete:
    post:
      operationId: completeLesson
      tags: [lessons]
      summary: Mark a lesson as completed for the current student
      parameters:
        - name: lessonId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LessonCompletion"
...
  /quizzes:
    get:
      operationId: listQuizzes
      tags: [quizzes]
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Quiz"
    post:
      operationId: createQuiz
      tags: [quizzes]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateQuizInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Quiz"
  /quizzes/{quizId}:
    get:
      operationId: getQuiz
      tags: [quizzes]
      parameters:
        - name: quizId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/QuizDetail"
    patch:
      operationId: updateQuiz
      tags: [quizzes]
      parameters:
        - name: quizId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateQuizInput"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Quiz"
    delete:
      operationId: deleteQuiz
      tags: [quizzes]
      parameters:
        - name: quizId
          in: path
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Deleted
  /quizzes/{quizId}/attempts:
    post:
      operationId: submitQuizAttempt
      tags: [quiz-attempts]
      parameters:
        - name: quizId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SubmitQuizAttemptInput"
      responses:
        "201":
          description: Scored
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/QuizAttemptResult"
...
components:
  schemas:
...
    Lesson:
      type: object
      required: [id, courseId, title, content, orderIndex, createdAt]
      properties:
        id:
          type: string
        courseId:
          type: string
        title:
          type: string
        content:
          type: string
        durationMinutes:
          type: integer
        orderIndex:
          type: integer
        createdAt:
          type: string
          format: date-time
    CreateLessonInput:
      type: object
      required: [title, content]
      properties:
        title:
          type: string
        content:
          type: string
        durationMinutes:
          type: integer
        orderIndex:
          type: integer
    UpdateLessonInput:
      type: object
      properties:
        title:
          type: string
        content:
          type: string
        durationMinutes:
          type: integer
        orderIndex:
          type: integer
    LessonCompletion:
      type: object
      required: [lessonId, completedAt]
      properties:
        lessonId:
          type: string
        completedAt:
          type: string
          format: date-time
...
    Quiz:
      type: object
      required: [id, courseId, courseTitle, title, description, questionCount, totalPoints, createdAt]
      properties:
        id:
          type: string
        courseId:
          type: string
        courseTitle:
          type: string
        title:
          type: string
        description:
          type: string
        questionCount:
          type: integer
        totalPoints:
          type: integer
        createdAt:
          type: string
          format: date-time
    QuizDetail:
      allOf:
        - $ref: "#/components/schemas/Quiz"
        - type: object
          required: [questions]
          properties:
            questions:
              type: array
              items:
                $ref: "#/components/schemas/QuizQuestion"
    QuizQuestion:
      type: object
      required: [id, prompt, options, points, orderIndex]
      properties:
        id:
          type: string
        prompt:
          type: string
        options:
          type: array
          items:
            type: string
        points:
          type: integer
        orderIndex:
          type: integer
    CreateQuizInput:
      type: object
      required: [courseId, title, description, questions]
      properties:
        courseId:
          type: string
        title:
          type: string
        description:
          type: string
        questions:
          type: array
          items:
            $ref: "#/components/schemas/CreateQuizQuestionInput"
    UpdateQuizInput:
      type: object
      properties:
        title:
          type: string
        description:
          type: string
        questions:
          type: array
          items:
            $ref: "#/components/schemas/CreateQuizQuestionInput"
    CreateQuizQuestionInput:
      type: object
      required: [prompt, options, correctIndex, points]
      properties:
        prompt:
          type: string
        options:
          type: array
          items:
            type: string
        correctIndex:
          type: integer
        points:
          type: integer
    SubmitQuizAttemptInput:
      type: object
      required: [answers]
      properties:
        answers:
          type: array
          items:
            type: object
            required: [questionId, selectedIndex]
            properties:
              questionId:
                type: string
              selectedIndex:
                type: integer
    QuizAttemptResult:
      type: object
      required: [id, quizId, score, totalPoints, percent, submittedAt, breakdown]
      properties:
        id:
          type: string
        quizId:
          type: string
        score:
          type: integer
        totalPoints:
          type: integer
        percent:
          type: integer
        submittedAt:
          type: string
          format: date-time
        breakdown:
          type: array
          items:
            type: object
            required: [questionId, prompt, selectedIndex, correctIndex, isCorrect]
            properties:
              questionId:
                type: string
              prompt:
                type: string
              selectedIndex:
                type: integer
              correctIndex:
                type: integer
              isCorrect:
                type: boolean
...