# Desktop/lumen/lib/api-zod/src/generated/types/createLessonInput.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Learning Management System API
 * OpenAPI spec version: 0.1.0
 */

export interface CreateLessonInput {
  title: string;
  content: string;
  durationMinutes?: number;
  orderIndex?: number;
}