# Desktop/lumen/lib/api-zod/src/generated/types/quiz.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Learning Management System API
 * OpenAPI spec version: 0.1.0
 */

export interface Quiz {
  id: string;
  courseId: string;
  courseTitle: string;
  title: string;
  description: string;
  questionCount: number;
  totalPoints: number;
  createdAt: Date;
}