# Desktop/lumen/lib/api-zod/src/generated/types/submitQuizAttemptInput.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Learning Management System API
 * OpenAPI spec version: 0.1.0
 */
import type { SubmitQuizAttemptInputAnswersItem } from "./submitQuizAttemptInputAnswersItem";

export interface SubmitQuizAttemptInput {
  answers: SubmitQuizAttemptInputAnswersItem[];
}